payment_service/requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
pydantic>=2.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3
PyJWT>=2.8
httpx>=0.25
pytest>=7.4
pytest-asyncio>=0.21
httpx>=0.25
